+++ lab4/lib/utils.h
@@ -3,5 +3,5 @@ void itoa(int num, char *s);
 int atoi(char* s, int base);
 int ls(char input[20], int readcontent);
 void print_welcome(int index);
-void dec_hex(int number, char *hexadecimal);
+void dec_hex(unsigned long number, char *hexadecimal);
 void boot_from_initramfs();
